+++ app/src/main/java/pm/gnosis/heimdall/ui/transactions/view/review/ReviewTransactionContract.kt
@@ -1,8 +1,9 @@
-package pm.gnosis.heimdall.ui.transactions.review
+package pm.gnosis.heimdall.ui.transactions.view.review
 
 import android.arch.lifecycle.ViewModel
 import io.reactivex.Observable
 import pm.gnosis.heimdall.data.repositories.TransactionData
+import pm.gnosis.heimdall.ui.transactions.view.TransactionInfoViewHolder
 import pm.gnosis.model.Solidity
 import pm.gnosis.models.Wei
 import pm.gnosis.svalinn.common.utils.Result
